+++ pkg/secrets/http/config.go
@@ -2,7 +2,13 @@ package http
 
 import (
 	"context"
+	"fmt"
+	"os"
+	"path/filepath"
 	"reflect"
+
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/zostay/ghost/pkg/config"
 	"github.com/zostay/ghost/pkg/plugin"
@@ -10,19 +16,31 @@ import (
 )
 
 // ConfigType is the type name for the HTTP secrets keeper.
-const ConfigType = "http"
+const (
+	ConfigType  = "http"
+	ServiceName = "ghost.keeper"
+)
 
 // Config is the configuration of the HTTP secrets keeper.
 type Config struct{}
 
 // Builder is the builder function for the HTTP secrets keeper.
-func Builder(_ context.Context, c any) (secrets.Keeper, error) {
+func Builder(ctx context.Context, c any) (secrets.Keeper, error) {
 	_, isGrpc := c.(*Config)
 	if !isGrpc {
 		return nil, plugin.ErrConfig
 	}
 
-	return NewClient(), nil
+	sock := MakeHttpServerSocketName()
+	clientConn, err := grpc.DialContext(ctx, "unix:"+sock,
+		grpc.WithTransportCredentials(insecure.NewCredentials()))
+	if err != nil {
+		return nil, err
+	}
+
+	client := NewKeeperClient(clientConn)
+
+	return NewClient(client), nil
 }
 
 func init() {
@@ -36,4 +54,10 @@ func init() {
 	}
 
 	plugin.Register(ConfigType, reflect.TypeOf(Config{}), Builder, nil, cmd)
+}
+
+func MakeHttpServerSocketName() string {
+	tmp := os.TempDir()
+	uid := os.Getuid()
+	return filepath.Join(tmp, fmt.Sprintf("%s.%d", ServiceName, uid))
 }