+++ pkg/secrets/keepass/config.go
@@ -2,7 +2,10 @@ package keepass
 
 import (
 	"context"
+	"os"
 	"reflect"
+
+	"github.com/mitchellh/go-homedir"
 
 	"github.com/zostay/ghost/pkg/config"
 	"github.com/zostay/ghost/pkg/plugin"
@@ -27,7 +30,12 @@ func Builder(_ context.Context, c any) (secrets.Keeper, error) {
 		return nil, plugin.ErrConfig
 	}
 
-	return NewKeepass(cfg.Path, cfg.Master)
+	path, err := homedir.Expand(os.ExpandEnv(cfg.Path))
+	if err != nil {
+		return nil, err
+	}
+
+	return NewKeepass(path, cfg.Master)
 }
 
 func init() {